OPENEUROPA-1383: Rework test classes structure.
